+++ UiLepra/Include/X11/UiX11Input.h
@@ -131,8 +131,8 @@ public:
 	virtual void ShowCursor();
 	virtual void HideCursor();
 
-	virtual double GetCursorX();
-	virtual double GetCursorY();
+	virtual float GetCursorX();
+	virtual float GetCursorY();
 
 	bool IsInitialized();
 
@@ -163,8 +163,8 @@ private:
 	int mScreenHeight;
 
 	// Mouse related stuff.
-	double mCursorX;
-	double mCursorY;
+	float mCursorX;
+	float mCursorY;
 
 	// Default devices.
 	InputDevice* mKeyboard;
